# specs/contracts/api/openapi.yaml
openapi: 3.1.0
info:
  title: muti-platform 只读 API v1
  description: 小学语数英个性化练习平台 · 只读查询 API。覆盖 item / item_version / item_template / gate_certificate。写入路径走
    src/core/content/writer.py 等服务层，本 API 不暴露。
  version: 1.0.0
paths:
  /items/{item_id}:
    get:
      tags:
      - items
      summary: 查询 item 身份与当前版本
      description: '返回 item 不变身份 + current_version_id 指向的版本内容（若有）.


        current_version 为 None 表示该 item 尚未发布任何版本。'
      operationId: get_item_items__item_id__get
      parameters:
      - name: item_id
        in: path
        required: true
        schema:
          type: string
          title: Item Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ItemDetailResponse'
        '404':
          description: item_id 不存在
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /item_versions/{item_version_id}:
    get:
      tags:
      - items
      summary: 查询 item_version 内容与谱系
      description: 返回 item_version 六大块 + 谱系 + 门证书引用（若已发布）.
      operationId: get_item_version_item_versions__item_version_id__get
      parameters:
      - name: item_version_id
        in: path
        required: true
        schema:
          type: string
          title: Item Version Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ItemVersionPydantic'
        '404':
          description: item_version_id 不存在
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /templates/{template_id}:
    get:
      tags:
      - items
      summary: 查询母题身份与当前版本
      description: 返回母题不变身份 + current_version_id 指向的母题版本（若有）.
      operationId: get_template_templates__template_id__get
      parameters:
      - name: template_id
        in: path
        required: true
        schema:
          type: string
          title: Template Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TemplateDetailResponse'
        '404':
          description: template_id 不存在
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /gate_certificates/{cert_id}:
    get:
      tags:
      - gate
      summary: 查询门证书与运行记录
      description: 返回门证书 + 关联的所有 gate_run（含 gate_verdict 明细）.
      operationId: get_gate_certificate_gate_certificates__cert_id__get
      parameters:
      - name: cert_id
        in: path
        required: true
        schema:
          type: string
          title: Cert Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GateCertificateRead'
        '404':
          description: cert_id 不存在
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /health:
    get:
      tags:
      - meta
      summary: 健康检查
      description: 健康检查端点（无需 DB 连接）.
      operationId: health_health_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
                title: Response Health Health Get
components:
  schemas:
    Content:
      properties:
        blocks:
          items:
            additionalProperties: true
            type: object
          type: array
          title: Blocks
      additionalProperties: true
      type: object
      required:
      - blocks
      title: Content
      description: '§2.2 content 块：题面语义 AST + 素材版本引用.


        为什么 blocks 用 list[dict] 而非强类型：题面 AST 因交互类型而异（选择题/

        填空题/拖拽题块结构各不相同），D4 注册表承载具体校验；本处保持 permissive

        让所有交互类型都能通过，是契约 §5 未对 content 强制 JSON Schema 的体现。'
    CorpusRef:
      properties:
        corpus_version_id:
          type: string
          title: Corpus Version Id
        digest:
          type: string
          title: Digest
      additionalProperties: false
      type: object
      required:
      - corpus_version_id
      - digest
      title: CorpusRef
      description: 语料库版本引用（lineage.corpus_refs 元素）。
    ErrorBindings:
      items:
        additionalProperties: true
        type: object
      type: array
      title: ErrorBindings
      description: '§2.2 error_bindings 块：选项/评分维度 → 错误类型 + 置信规则（R-Q-06/07）.


        为什么用 RootModel：error_bindings 在 JSONB 中顶层是数组（list[dict]），

        每个元素是一个错误绑定；用 RootModel 让 Pydantic 模型直接对应数组，

        而非 {bindings: [...]} 的包装对象，与 DB 存储结构对齐。

        为什么元素是 permissive dict：契约 §5 未对 error_bindings 单元素结构

        强制 JSON Schema，R-Q-06/07 的具体结构由错误类型注册表承载。'
    GateCertificateRead:
      properties:
        cert_id:
          type: string
          title: Cert Id
        artifact_ref:
          type: string
          title: Artifact Ref
        cert_type:
          type: string
          title: Cert Type
        policy_version:
          type: string
          title: Policy Version
        issued_by:
          type: string
          title: Issued By
        issued_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Issued At
        created_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Created At
        runs:
          items:
            $ref: '#/components/schemas/GateRunRead'
          type: array
          title: Runs
      additionalProperties: false
      type: object
      required:
      - cert_id
      - artifact_ref
      - cert_type
      - policy_version
      - issued_by
      title: GateCertificateRead
      description: 门证书 + 关联 runs（GET /gate_certificates/{cert_id} 响应）.
    GateRunRead:
      properties:
        run_id:
          type: string
          title: Run Id
        certificate_id:
          type: string
          title: Certificate Id
        policy_version:
          type: string
          title: Policy Version
        validator_id:
          type: string
          title: Validator Id
        validator_version:
          type: string
          title: Validator Version
        verdict:
          type: string
          title: Verdict
        evidence:
          additionalProperties: true
          type: object
          title: Evidence
        confidence:
          type: string
          pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          title: Confidence
        cost_ms:
          type: integer
          title: Cost Ms
        cost_tokens:
          type: integer
          title: Cost Tokens
        run_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Run At
        created_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Created At
        verdicts:
          items:
            $ref: '#/components/schemas/GateVerdictRead'
          type: array
          title: Verdicts
      additionalProperties: false
      type: object
      required:
      - run_id
      - certificate_id
      - policy_version
      - validator_id
      - validator_version
      - verdict
      - evidence
      - confidence
      - cost_ms
      - cost_tokens
      title: GateRunRead
      description: 单次验证器运行记录（读取响应）.
    GateVerdictRead:
      properties:
        verdict_id:
          type: integer
          title: Verdict Id
        run_id:
          type: string
          title: Run Id
        detail:
          additionalProperties: true
          type: object
          title: Detail
        created_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Created At
      additionalProperties: false
      type: object
      required:
      - verdict_id
      - run_id
      - detail
      title: GateVerdictRead
      description: 门判定明细（读取响应）.
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    InteractionRef:
      properties:
        interaction_id:
          type: string
          title: Interaction Id
        interaction_params:
          additionalProperties: true
          type: object
          title: Interaction Params
      additionalProperties: false
      type: object
      required:
      - interaction_id
      - interaction_params
      title: InteractionRef
      description: '§2.2 interaction_ref 块：交互类型 + 交互参数.


        interaction_id 必须在 registries/interaction.yaml 注册（D4）；

        interaction_params 的具体结构由交互类型决定，本处保持 dict 不细化。'
    ItemDetailResponse:
      properties:
        item_id:
          type: string
          title: Item Id
        pack_id:
          type: string
          title: Pack Id
        tier:
          type: string
          enum:
          - A
          - B
          - C
          - D
          title: Tier
        template_version_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Template Version Id
        current_version_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Current Version Id
        created_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Created At
        current_version:
          anyOf:
          - $ref: '#/components/schemas/ItemVersionPydantic'
          - type: 'null'
      additionalProperties: false
      type: object
      required:
      - item_id
      - pack_id
      - tier
      title: ItemDetailResponse
      description: 'GET /items/{item_id} 响应：item 身份 + current_version（若有）.


        current_version 为 None 表示该 item 尚无已发布版本（current_version_id IS NULL）。'
    ItemTemplateVersionPydantic:
      properties:
        template_version_id:
          type: string
          title: Template Version Id
        template_id:
          type: string
          title: Template Id
        dsl_version:
          type: string
          title: Dsl Version
        spec:
          additionalProperties: true
          type: object
          title: Spec
        status:
          type: string
          enum:
          - draft
          - published
          - retired
          title: Status
        created_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Created At
      additionalProperties: false
      type: object
      required:
      - template_version_id
      - template_id
      - dsl_version
      - spec
      - status
      title: ItemTemplateVersionPydantic
      description: 'ItemTemplateVersion 的 Pydantic 表示.


        spec 字段保持 dict 不细化：母题 DSL 结构随 DSL 版本演化，本处不强约束。'
    ItemVersionPydantic:
      properties:
        item_version_id:
          type: string
          title: Item Version Id
        item_id:
          type: string
          title: Item Id
        status:
          type: string
          enum:
          - draft
          - quarantined
          - published
          - retired
          title: Status
        objective:
          $ref: '#/components/schemas/Objective'
        interaction_ref:
          $ref: '#/components/schemas/InteractionRef'
        content:
          $ref: '#/components/schemas/Content'
        scoring_ref:
          $ref: '#/components/schemas/ScoringRef'
        error_bindings:
          $ref: '#/components/schemas/ErrorBindings'
        lineage:
          $ref: '#/components/schemas/Lineage'
        rendered_snapshot:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Rendered Snapshot
        gate_certificate_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Gate Certificate Id
        published_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Published At
        retired_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Retired At
        created_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Created At
      additionalProperties: false
      type: object
      required:
      - item_version_id
      - item_id
      - status
      - objective
      - interaction_ref
      - content
      - scoring_ref
      - error_bindings
      - lineage
      title: ItemVersionPydantic
      description: 'ItemVersion 的 Pydantic 表示（聚合六大块强类型子模型）.


        用于 API 入参校验/序列化；created_at 等服务端填字段保持 Optional。'
    KpRef:
      properties:
        dimension:
          type: string
          title: Dimension
        code:
          type: string
          title: Code
      additionalProperties: false
      type: object
      required:
      - dimension
      - code
      title: KpRef
      description: '知识点引用（契约 §2.2.1 kp_set 元素）.


        dimension：知识维度（''kp'' 等，供图谱分面查询）。

        code：知识点编码（如 ''math.nal.decimal.compare''）。'
    Lineage:
      properties:
        tier:
          type: string
          enum:
          - A
          - B
          - C
          - D
          title: Tier
        pipeline:
          $ref: '#/components/schemas/Pipeline'
        template_version_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Template Version Id
        params:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Params
        seed:
          anyOf:
          - type: integer
          - type: 'null'
          title: Seed
        corpus_refs:
          anyOf:
          - items:
              $ref: '#/components/schemas/CorpusRef'
            type: array
          - type: 'null'
          title: Corpus Refs
        ai_ledger_refs:
          anyOf:
          - items:
              type: string
            type: array
          - type: 'null'
          title: Ai Ledger Refs
        signed_by:
          type: string
          title: Signed By
        signed_at:
          type: string
          title: Signed At
      additionalProperties: false
      type: object
      required:
      - tier
      - pipeline
      - signed_by
      - signed_at
      title: Lineage
      description: '§2.2.2 lineage 块：生产谱系（R-Q-22）.


        对应契约 §5.2 机器可校验 JSON Schema；

        required = [tier, pipeline, signed_by, signed_at]；

        tier ∈ {A,B} 时 template_version_id 与 params 必填（应用层校验，schema 不强制）；

        tier ∈ {C,D} 且经 AI 起草时 ai_ledger_refs 非空（同上）。'
    Objective:
      properties:
        kp_set:
          items:
            $ref: '#/components/schemas/KpRef'
          type: array
          title: Kp Set
        kp_set_mode:
          type: string
          enum:
          - single
          - all_required
          - compensatory
          title: Kp Set Mode
        cognitive_level:
          type: string
          enum:
          - remember
          - understand
          - apply
          - analyze
          - evaluate
          - create
          title: Cognitive Level
        gradeband:
          type: string
          enum:
          - L
          - M
          - H
          title: Gradeband
        graph_release:
          type: string
          title: Graph Release
        steps:
          anyOf:
          - items:
              $ref: '#/components/schemas/StepRef'
            type: array
          - type: 'null'
          title: Steps
      additionalProperties: false
      type: object
      required:
      - kp_set
      - kp_set_mode
      - cognitive_level
      - gradeband
      - graph_release
      title: Objective
      description: '§2.2.1 objective 块：知识标注集 + 认知层级 + 学段 + 图谱版本.


        对应契约 §5.1 机器可校验 JSON Schema；required 字段缺一即校验失败。'
    Pipeline:
      properties:
        id:
          type: string
          title: Id
        version:
          type: string
          title: Version
      additionalProperties: false
      type: object
      required:
      - id
      - version
      title: Pipeline
      description: 生产线标识（lineage.pipeline）。
    ScoringRef:
      properties:
        scorer_id:
          type: string
          title: Scorer Id
        scorer_params:
          additionalProperties: true
          type: object
          title: Scorer Params
      additionalProperties: false
      type: object
      required:
      - scorer_id
      - scorer_params
      title: ScoringRef
      description: '§2.2 scoring_ref 块：评分器 + 评分参数.


        scorer_id 必须在 registries/scorer.yaml 注册（D4）；

        scorer_params 的结构由评分器决定，本处保持 dict 不细化。'
    StepRef:
      properties:
        step_id:
          type: string
          title: Step Id
        kp:
          items:
            type: string
          type: array
          title: Kp
      additionalProperties: false
      type: object
      required:
      - step_id
      - kp
      title: StepRef
      description: 分步过程题的步骤级标注（R-Q-15）.
    TemplateDetailResponse:
      properties:
        template_id:
          type: string
          title: Template Id
        pack_id:
          type: string
          title: Pack Id
        current_version_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Current Version Id
        created_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Created At
        current_version:
          anyOf:
          - $ref: '#/components/schemas/ItemTemplateVersionPydantic'
          - type: 'null'
      additionalProperties: false
      type: object
      required:
      - template_id
      - pack_id
      title: TemplateDetailResponse
      description: GET /templates/{template_id} 响应：母题身份 + current_version（若有）.
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError